main page and countries page
countries page was the main page. This is fixed
