+++ React/src/app/countries/page.module.css
@@ -1,0 +1,140 @@
+#root {
+  font-family: Arial, sans-serif;
+  background-color: whitesmoke;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  margin: 0;
+}
+
+:global(body) {
+  font-family: Arial, sans-serif;
+  background-color: whitesmoke;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  
+  margin: 0;
+}
+
+.container {
+  width: 100%;
+  text-align: center;
+  overflow-y: auto;
+}
+
+:global(h1) {
+  color: #333;
+  font-size: 2em;
+}
+
+.flag_grid {
+  width: 80%;
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  gap: 20px;
+  max-height: 70vh;
+  overflow-y: auto;
+  place-items: center;
+  margin: 50px auto 0 auto;
+}
+
+.flag_card { 
+  width: 200px !important;
+  padding: 10px;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  text-align: center;
+  transition: transform 0.2s;
+  transition: background-color 0.6s;
+}
+
+.flag_card:hover {
+  transform: translateY(-5px);
+  background-color: aqua;
+}
+
+.flag_card img {
+  height: 70px !important;
+  border-radius: 4px;
+  margin-bottom: 10px;
+}
+
+.flag_card p {
+  font-size: 1em;
+  color: #333;
+  margin: 0;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.search_bar {
+  position: relative;
+  display: inline-block;
+  padding-bottom: 10px;
+}
+
+.search_bar input[type="text"] {
+  padding-left: 5px;
+}
+
+.search_icon {
+  padding-bottom: 10px;
+  position: absolute;
+  float: right;
+  right: 8px;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 16px; 
+  height: 16px; 
+  transition: width 0.3s ease, height 0.3s ease;
+  box-sizing: content-box !important;
+}
+
+.search_icon:hover {
+  width: 20px;
+  height: 20px;
+
+}
+
+.pagination {
+  display: flex;
+  position: absolute;
+  gap: 5px;
+  align-items: center;
+  justify-content: center;
+  margin-top: 20px;
+  bottom: 0;
+  font-size: 1.2em;
+  width: 100%;
+  padding-bottom: 30px;
+}
+
+:global(button) {
+  padding: 5px 10px;
+  cursor: pointer;
+}
+
+:global(button:disabled) {
+  cursor: not-allowed;
+  opacity: 0.5;
+}
+
+:global(button.active) {
+  font-weight: bold;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+}
+
+.top {
+  position: absolute;
+  align-items: center;
+  justify-content: center;
+  margin-top: 20px;
+  top: 0;
+  font-size: 1.2em;
+  width: 100%;
+}
